+++ pages/boards/components/BestBoardList/BestBoardCard/index.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import React from "react";
 import Image from "next/image";
 import styles from "../styles.module.scss";
